backend/requirements.txt
firebase-admin==6.0.1
google-cloud-firestore==2.11.0
google-cloud-storage==2.13.0
flask==2.2.5
google-cloud-secret-manager>=2.0.0
Pillow==9.5.0
